refactor: move private methods to db access module
